[rename] domain 모듈 패키지 구조 수정 등 (#20)
