+++ domain/src/main/java/com/freetreechair/domain/auth/login/model/DomainLoginRequest.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.auth.login.model
+
+data class DomainLoginRequest(
+    val platform: String,
+    val socialToken: String
+)
