+++ domain/src/main/java/com/freetreechair/domain/auth/login/model/DomainLoginResponse.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.auth.login.model
+
+data class DomainLoginResponse(
+    val isNewUser: Boolean,
+    val accessToken: String
+)
