+++ domain/src/main/java/com/freetreechair/domain/auth/login/usecase/GetAccessTokenUseCase.kt
@@ -1,0 +1,11 @@
+package com.freetreechair.domain.auth.login.usecase
+
+import com.freetreechair.domain.auth.login.repository.LoginRepository
+
+class GetAccessTokenUseCase(
+    private val repository: LoginRepository
+) {
+    operator fun invoke(): String {
+        return repository.getAccessToken()
+    }
+}
