+++ domain/src/main/java/com/freetreechair/domain/auth/login/usecase/LoginUseCases.kt
@@ -1,0 +1,7 @@
+package com.freetreechair.domain.auth.login.usecase
+
+data class LoginUseCases(
+    val getAccessTokenUseCase: GetAccessTokenUseCase,
+    val saveAccessTokenUseCase: SaveAccessTokenUseCase,
+    val makeLoginRequestUseCase: MakeLoginRequestUseCase
+)
