+++ domain/src/main/java/com/freetreechair/domain/auth/login/usecase/SaveAccessTokenUseCase.kt
@@ -1,0 +1,11 @@
+package com.freetreechair.domain.auth.login.usecase
+
+import com.freetreechair.domain.auth.login.repository.LoginRepository
+
+class SaveAccessTokenUseCase(
+    private val repository: LoginRepository
+) {
+    operator fun invoke(accessToken: String) {
+        repository.saveAccessToken(accessToken)
+    }
+}
