--- domain/src/main/java/com/freetreechair/domain/disgust/model/UIDisgust.kt
@@ -1,8 +1,0 @@
-package com.freetreechair.domain.disgust.model
-
-data class UIDisgust(
-    val id: Int,
-    val name: String,
-    val imageUrl: String,
-    val isChecked: Boolean = false
-)
