--- domain/src/main/java/com/freetreechair/domain/disgust/usecase/DisgustUseCases.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.disgust.usecase
-
-data class DisgustUseCases(
-    val fetchDisgustUseCase: FetchDisgustUseCase,
-    val saveDisgustUseCase: SaveDisgustUseCase
-)
