--- domain/src/main/java/com/freetreechair/domain/disgust/usecase/SaveDisgustUseCase.kt
@@ -1,11 +1,0 @@
-package com.freetreechair.domain.disgust.usecase
-
-import com.freetreechair.domain.disgust.repository.DisgustRepository
-
-class SaveDisgustUseCase(
-    private val repository: DisgustRepository
-) {
-    operator fun invoke(disgusts: String) {
-        repository.saveDisgusts(disgusts)
-    }
-}
