--- domain/src/main/java/com/freetreechair/domain/evaluate/repository/EvaluateRepository.kt
@@ -1,8 +1,0 @@
-package com.freetreechair.domain.evaluate.repository
-
-import com.freetreechair.domain.evaluate.model.UIRestaurant
-
-interface EvaluateRepository {
-    suspend fun fetchRestaurants(query: String?): Result<List<UIRestaurant>>
-    fun saveEvaluates(evaluates: String)
-}
