--- domain/src/main/java/com/freetreechair/domain/evaluate/usecase/EvaluateUseCases.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.evaluate.usecase
-
-data class EvaluateUseCases(
-    val fetchRestaurantUseCase: FetchRestaurantUseCase,
-    val saveEvaluateUseCase: SaveEvaluateUseCase
-)
