--- domain/src/main/java/com/freetreechair/domain/evaluate/usecase/FetchRestaurantUseCase.kt
@@ -1,13 +1,0 @@
-package com.freetreechair.domain.evaluate.usecase
-
-import com.freetreechair.domain.evaluate.model.UIRestaurant
-import com.freetreechair.domain.evaluate.repository.EvaluateRepository
-
-class FetchRestaurantUseCase(
-    private val repository: EvaluateRepository
-) {
-    suspend operator fun invoke(query: String?): Result<List<UIRestaurant>> {
-        return repository.fetchRestaurants(query)
-    }
-}
-
