--- domain/src/main/java/com/freetreechair/domain/evaluate/usecase/SaveEvaluateUseCase.kt
@@ -1,11 +1,0 @@
-package com.freetreechair.domain.evaluate.usecase
-
-import com.freetreechair.domain.evaluate.repository.EvaluateRepository
-
-class SaveEvaluateUseCase(
-    private val repository: EvaluateRepository
-) {
-    operator fun invoke(evaluates: String) {
-        repository.saveEvaluates(evaluates)
-    }
-}
