--- domain/src/main/java/com/freetreechair/domain/login/model/DomainLoginRequest.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.login.model
-
-data class DomainLoginRequest(
-    val platform: String,
-    val socialToken: String
-)
