--- domain/src/main/java/com/freetreechair/domain/login/model/DomainLoginResponse.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.login.model
-
-data class DomainLoginResponse(
-    val isNewUser: Boolean,
-    val accessToken: String
-)
