--- domain/src/main/java/com/freetreechair/domain/login/usecase/GetAccessTokenUseCase.kt
@@ -1,11 +1,0 @@
-package com.freetreechair.domain.login.usecase
-
-import com.freetreechair.domain.login.repository.LoginRepository
-
-class GetAccessTokenUseCase(
-    private val repository: LoginRepository
-) {
-    operator fun invoke(): String {
-        return repository.getAccessToken()
-    }
-}
