--- domain/src/main/java/com/freetreechair/domain/login/usecase/LoginUseCases.kt
@@ -1,7 +1,0 @@
-package com.freetreechair.domain.login.usecase
-
-data class LoginUseCases(
-    val getAccessTokenUseCase: GetAccessTokenUseCase,
-    val saveAccessTokenUseCase: SaveAccessTokenUseCase,
-    val makeLoginRequestUseCase: MakeLoginRequestUseCase
-)
