--- domain/src/main/java/com/freetreechair/domain/nickname/repository/NicknameRepository.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.nickname.repository
-
-interface NicknameRepository {
-    suspend fun checkIsNicknameDuplicated(nickname: String): Result<Boolean>
-    fun saveNickname(nickname: String)
-}
