--- domain/src/main/java/com/freetreechair/domain/nickname/usecase/NicknameUseCases.kt
@@ -1,6 +1,0 @@
-package com.freetreechair.domain.nickname.usecase
-
-data class NicknameUseCases(
-    val checkIsNicknameDuplicatedUseCase: CheckIsNicknameDuplicatedUseCase,
-    val saveNicknameUseCase: SaveNicknameUseCase
-)
