+++ domain/src/main/java/com/freetreechair/domain/signup/complete/model/DomainSignUpResponse.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.complete.model
+
+data class DomainSignUpResponse(
+    val userId: Int,
+    val accessToken: String
+)
