+++ domain/src/main/java/com/freetreechair/domain/signup/complete/usecase/CompleteUseCases.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.complete.usecase
+
+data class CompleteUseCases(
+    val getUserIdentificationUseCase: GetUserIdentificationUseCase,
+    val makeSignUpRequestUseCase: MakeSignUpRequestUseCase
+)
