+++ domain/src/main/java/com/freetreechair/domain/signup/complete/usecase/GetUserIdentificationUseCase.kt
@@ -1,0 +1,11 @@
+package com.freetreechair.domain.signup.complete.usecase
+
+import com.freetreechair.domain.signup.complete.repository.CompleteRepository
+
+class GetUserIdentificationUseCase(
+    private val repository: CompleteRepository
+) {
+    operator fun invoke(): HashMap<String, String> {
+        return repository.getUserIdentification()
+    }
+}
