+++ domain/src/main/java/com/freetreechair/domain/signup/disgust/model/UIDisgust.kt
@@ -1,0 +1,8 @@
+package com.freetreechair.domain.signup.disgust.model
+
+data class UIDisgust(
+    val id: Int,
+    val name: String,
+    val imageUrl: String,
+    val isChecked: Boolean = false
+)
