+++ domain/src/main/java/com/freetreechair/domain/signup/disgust/usecase/DisgustUseCases.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.disgust.usecase
+
+data class DisgustUseCases(
+    val fetchDisgustUseCase: FetchDisgustUseCase,
+    val saveDisgustUseCase: SaveDisgustUseCase
+)
