+++ domain/src/main/java/com/freetreechair/domain/signup/evaluate/usecase/EvaluateUseCases.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.evaluate.usecase
+
+data class EvaluateUseCases(
+    val fetchRestaurantUseCase: FetchRestaurantUseCase,
+    val saveEvaluateUseCase: SaveEvaluateUseCase
+)
