+++ domain/src/main/java/com/freetreechair/domain/signup/nickname/repository/NicknameRepository.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.nickname.repository
+
+interface NicknameRepository {
+    suspend fun checkIsNicknameDuplicated(nickname: String): Result<Boolean>
+    fun saveNickname(nickname: String)
+}
