+++ domain/src/main/java/com/freetreechair/domain/signup/nickname/usecase/CheckIsNicknameDuplicatedUseCase.kt
@@ -1,0 +1,11 @@
+package com.freetreechair.domain.signup.nickname.usecase
+
+import com.freetreechair.domain.signup.nickname.repository.NicknameRepository
+
+class CheckIsNicknameDuplicatedUseCase(
+    private val repository: NicknameRepository
+) {
+    suspend operator fun invoke(nickname: String): Result<Boolean> {
+        return repository.checkIsNicknameDuplicated(nickname)
+    }
+}
