+++ domain/src/main/java/com/freetreechair/domain/signup/nickname/usecase/NicknameUseCases.kt
@@ -1,0 +1,6 @@
+package com.freetreechair.domain.signup.nickname.usecase
+
+data class NicknameUseCases(
+    val checkIsNicknameDuplicatedUseCase: CheckIsNicknameDuplicatedUseCase,
+    val saveNicknameUseCase: SaveNicknameUseCase
+)
