+++ domain/src/main/java/com/freetreechair/domain/signup/taste/repository/TasteRepository.kt
@@ -1,0 +1,5 @@
+package com.freetreechair.domain.signup.taste.repository
+
+interface TasteRepository {
+    fun saveTastes(tastes: String)
+}
