+++ domain/src/main/java/com/freetreechair/domain/signup/taste/usecase/TasteUseCases.kt
@@ -1,0 +1,5 @@
+package com.freetreechair.domain.signup.taste.usecase
+
+data class TasteUseCases(
+    val saveTasteUseCase: SaveTasteUseCase
+)
