--- domain/src/main/java/com/freetreechair/domain/taste/repository/TasteRepository.kt
@@ -1,5 +1,0 @@
-package com.freetreechair.domain.taste.repository
-
-interface TasteRepository {
-    fun saveTastes(tastes: String)
-}
