--- domain/src/main/java/com/freetreechair/domain/taste/usecase/SaveTasteUseCase.kt
@@ -1,11 +1,0 @@
-package com.freetreechair.domain.taste.usecase
-
-import com.freetreechair.domain.taste.repository.TasteRepository
-
-class SaveTasteUseCase(
-    private val repository: TasteRepository
-) {
-    operator fun invoke(tastes: String) {
-        repository.saveTastes(tastes)
-    }
-}
