--- domain/src/main/java/com/freetreechair/domain/taste/usecase/TasteUseCases.kt
@@ -1,5 +1,0 @@
-package com.freetreechair.domain.taste.usecase
-
-data class TasteUseCases(
-    val saveTasteUseCase: SaveTasteUseCase
-)
